Add lock when notify_one in ActorGroup
To avoid the main thread being blocked when waiting for actors to be done.
Without the lock, after the main thread checks the waiting condition,
the last actor may update the waiting condition and send the notify_one signal,
before the main thread actually starts to wait. In this case, the main thread
will wait forever as the last actor was done and the main thread missed the signal.